replaces other confirm boxes with hooks

+++ packages/cms/src/newLayout/components/CogMenu.jsx
@@ -1,11 +1,14 @@
 /* react */
-import React, {useState} from "react";
+import React from "react";
 import {useDispatch} from "react-redux";
-import {Modal, Menu, Button, Group} from "@mantine/core";
+import {Menu} from "@mantine/core";
 import {HiOutlineTrash, HiOutlineEye} from "react-icons/hi";
 
 /* redux */
 import {deleteEntity, deleteProfile} from "../../actions/profiles";
+
+/* hooks */
+import {useConfirmationDialog} from "../hooks/interactions/ConfirmationDialog";
 
 /* enums */
 import {ENTITY_TYPES, ENTITY_PRETTY_NAMES} from "../../utils/consts/cms";
@@ -18,30 +21,27 @@ function CogMenu({type, id, control}) {
   const PRETTY_NAME = ENTITY_PRETTY_NAMES[type] || "Entity";
 
   const dispatch = useDispatch();
-
-  const [opened, setOpened] = useState(false);
+  const {getConfirmation} = useConfirmationDialog();
 
   const toggleVisibility = () => {
     console.log("todo1.0 vis switch");
   };
 
-  const maybeDelete = () => setOpened(true);
-  const onClose = () => setOpened(false);
+  const maybeDelete = async() => {
+    const confirmed = await getConfirmation({
+      title: "Are you sure?",
+      message: `Delete this ${PRETTY_NAME} and all its contents? This action cannot be undone.`,
+      confirmText: "Yes, Delete it."
+    });
 
-  const onDelete = () => {
-    onClose();
-    if (type === ENTITY_TYPES.PROFILE) {
-      dispatch(deleteProfile(id));
+    if (confirmed) {
+      if (type === ENTITY_TYPES.PROFILE) {
+        dispatch(deleteProfile(id));
+      }
+      else {
+        dispatch(deleteEntity(type, {id}));
+      }
     }
-    else {
-      dispatch(deleteEntity(type, {id}));
-    }
-  };
-
-  const modalProps = {
-    opened,
-    title: "Are you sure?",
-    onClose
   };
 
   return (
@@ -51,13 +51,6 @@ function CogMenu({type, id, control}) {
         <Menu.Item icon={<HiOutlineEye />} onClick={toggleVisibility}>Visible</Menu.Item>
         <Menu.Item icon={<HiOutlineTrash />} onClick={maybeDelete}>Delete</Menu.Item>
       </Menu>
-      <Modal {...modalProps} key="modal">
-        Are you sure you want to delete this {PRETTY_NAME} and all its children? This action cannot be undone.
-        <Group position="right" style={{marginTop: 10}}>
-          <Button onClick={onClose}>Cancel</Button>
-          <Button color="red" onClick={onDelete}>Yes, Delete it.</Button>
-        </Group>
-      </Modal>
     </React.Fragment>
   );
