+++ packages/cms/src/newLayout/components/SectionMenu.jsx
@@ -1,12 +1,14 @@
 /* react */
-import React, {useState} from "react";
+import React from "react";
 import {useDispatch} from "react-redux";
-import {Modal, Menu, Button, Group} from "@mantine/core";
+import {Menu} from "@mantine/core";
 import {HiOutlineTrash} from "react-icons/hi";
-import {MdOutlineAccountTree} from "react-icons/md";
 
 /* redux */
 import {deleteEntity} from "../../actions/profiles";
+
+/* hooks */
+import {useConfirmationDialog} from "../hooks/interactions/ConfirmationDialog";
 
 /* enums */
 import {ENTITY_TYPES} from "../../utils/consts/cms";
@@ -14,24 +16,19 @@ import {ENTITY_TYPES} from "../../utils/consts/cms";
 /**
  *
  */
-function SectionMenu({id, control, toggleDependencies, showDependencies}) {
+function SectionMenu({id, control}) {
 
   const dispatch = useDispatch();
+  const {getConfirmation} = useConfirmationDialog();
 
-  const [opened, setOpened] = useState(false);
+  const maybeDelete = async() => {
+    const confirmed = await getConfirmation({
+      title: "Are you sure?",
+      message: "Are you sure you want to delete this Section and all its children? This action cannot be undone.",
+      confirmText: "Yes, Delete it."
+    });
 
-  const maybeDelete = () => setOpened(true);
-  const onClose = () => setOpened(false);
-
-  const onDelete = () => {
-    onClose();
-    dispatch(deleteEntity(ENTITY_TYPES.SECTION, {id}));
-  };
-
-  const modalProps = {
-    opened,
-    title: "Are you sure?",
-    onClose
+    if (confirmed) dispatch(deleteEntity(ENTITY_TYPES.SECTION, {id}));
   };
 
   return (
@@ -40,13 +37,6 @@ function SectionMenu({id, control, toggleDependencies, showDependencies}) {
         <Menu.Label>Settings</Menu.Label>
         <Menu.Item icon={<HiOutlineTrash />} onClick={maybeDelete}>Delete</Menu.Item>
       </Menu>
-      <Modal {...modalProps} key="modal">
-        Are you sure you want to delete this Section and all its children? This action cannot be undone.
-        <Group position="right" style={{marginTop: 10}}>
-          <Button onClick={onClose}>Cancel</Button>
-          <Button color="red" onClick={onDelete}>Yes, Delete it.</Button>
-        </Group>
-      </Modal>
     </React.Fragment>
   );
